requirements.txt
mcp[cli]
Pillow
cairosvg
spidev
RPi.GPIO
gpiozero
lgpio
